utils/requirements.txt
CairoSVG==2.7.0
Pillow==10.0.0
